refactor: sort la méthode renderSimulationResultsSmsUrl du followup vers sms-service.ts
